chore(web-app): replace duplicate RoleDropdown with shared RoleFilter component

fix: comments
